Working callbacks/power on and off buttons. Disabled AUTO_SHUTDOWN of PSU.
Callbacks from the WebServer library now include a pointer to a given object (usually server.cpp instance). There is also a power button on the webpage now.

Disabled Auto Shutdown of PSU in order to use the webpage on/off buttons. Best solution I have for now. It could probably be fixed to work with both in the future fairly easily.