Working communication between Pi and Photon
No longer attempting SPI. Using one wire for data, one as a 'data received' by the Photon and one as a 'data ready' from the Pi. The Pi puts data on the data line, toggles the value of the data ready line. When the Photon sees this line toggle it reads the data and toggles its data received line. The two go back and forth until all 8 bits are sent.